requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
backcall==0.2.0
batch-face==1.5.4
bce-python-sdk==0.9.56
bcrypt==5.0.0
beautifulsoup4==4.14.3
bleach==6.3.0
cachetools==6.2.4
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
contourpy==1.3.3
cryptography==43.0.3
cssselect==1.3.0
cssselect2==0.8.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.2.3
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
dropbox==12.0.2
easyocr==1.7.2
einops==0.8.1
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.1
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
ftfy==6.3.1
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-genai==1.56.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.2.3
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
imagesize==1.4.1
invoke==2.2.1
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.80
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langsmith==0.5.0
lazy_loader==0.4
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.8
matplotlib-inline==0.2.1
mistune==3.1.4
modelscope==1.33.0
more-itertools==10.8.0
mpmath==1.3.0
mss==10.1.0
multidict==6.7.0
mypy_extensions==1.1.0
nbclient==0.10.3
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.6.1
ninja==1.13.0
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauth2client==4.1.3
openai==2.14.0
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.11.0.86
opencv-transforms==0.0.6
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.5
packaging==25.0
paddleocr==3.3.2
paddlepaddle==3.2.0
paddlex==3.3.12
pandas==2.3.3
pandocfilters==1.5.1
paramiko==4.0.0
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.0.0
pipreqs==0.5.0
platformdirs==4.5.1
ply==3.11
premailer==3.10.0
prettytable==3.17.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyNaCl==1.6.1
pyOpenSSL==24.2.1
pyparsing==3.2.5
pypdfium2==5.2.0
PyQt6==6.10.1
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.3
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.8.0
scipy==1.16.3
scp==0.15.0
sentencepiece==0.2.1
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sixdrepnet==0.1.6
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
stack-data==0.6.3
stone==3.3.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.12.12
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
triton==3.5.1
typer-slim==0.20.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.2
uuid_utils==0.12.0
vidgear==0.3.4
wcwidth==0.2.14
webencodings==0.5.1
websockets==15.0.1
yarg==0.1.9
yarl==1.22.0
yt-dlp==2026.3.17
zstandard==0.25.0
